+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 12.103.0
 
 * `[options]` Fixed panic when parsing unsupported option with value passed with equal sign (`=`)
+* `[options]` Code refactoring
 
 ### 12.102.0